End of video 6
Added intersection detection for sphere, added function to find closest
object to the camera so that we can later pick that object color for the
pixel
It still is missing some OOP features such as functions and variables
that are present in Object subclasses but that are not present in
Object.
Also added the "ray generator" that generates rays in a "pyramidal" way
(not sure about that).