Started adding scene files
Also: Output files now have date and time and are stored in folder
'Resultados'.

Scene file parsing has been added for image resolution, AA threshold,
camera position and direction and for the creation of spheres. It lacks
a Color library or something like that, to make it easier to add color
to objects.
It should be implemented in a .txt file and there should be a map of
colors, so that users can specify colors like "orange" to objects.

Also Todo: verify the little black dots that appear when a cube is
rendered